refactor(api): remove insertOrganizationUserWithRoleAdmin from test-helper.js
